// src/webparts/employeePerformanceApp/models/IEmployeeInformation.ts
interface IEmployeeInformation {
  userPrincipalName: string;
  birthday: string;
  vehicule?: string;
  emergencyContacts: string;
  rewardPoints: number;
  hobbies: string;
  music: string;
  interests: string;
  food: string;
  blogs: string;
  facebook: string;
  twitter: string;
  linkedIn: string;
  officeCubicle: string;
  isHomeOffice: boolean;
}

export default IEmployeeInformation;